update the register token files packages to auth
