+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/data/event/EventRepository.kt
@@ -7,9 +7,7 @@ import com.wire.kalium.logic.data.conversation.ClientId
 import com.wire.kalium.logic.di.MapperProvider
 import com.wire.kalium.logic.functional.Either
 import com.wire.kalium.logic.functional.suspending
-import com.wire.kalium.logic.wrapApiRequest
 import com.wire.kalium.network.api.notification.NotificationApi
-import com.wire.kalium.network.api.notification.pushToken.PushTokenRequestBody
 import com.wire.kalium.network.api.notification.NotificationResponse
 import com.wire.kalium.network.utils.NetworkResponse
 import com.wire.kalium.network.utils.isSuccessful
@@ -28,7 +26,6 @@ import kotlin.coroutines.coroutineContext
 interface EventRepository {
     suspend fun events(): Flow<Either<CoreFailure, Event>>
     suspend fun updateLastProcessedEventId(eventId: String)
-    suspend fun registerToken(body: PushTokenRequestBody): Either<NetworkFailure, Unit>
 }
 
 class EventDataSource(
@@ -88,10 +85,6 @@ class EventDataSource(
 
     override suspend fun updateLastProcessedEventId(eventId: String) {
         eventInfoStorage.lastProcessedId = eventId
-    }
-
-    override suspend fun registerToken(body: PushTokenRequestBody): Either<NetworkFailure, Unit> = wrapApiRequest {
-        notificationApi.registerToken(body)
     }
 
     private suspend fun getNextPendingEventsPage(
