+++ logic/src/commonTest/kotlin/com/wire/kalium/logic/feature/session/PushTokenUseCaseTest.kt
@@ -1,10 +1,10 @@
 package com.wire.kalium.logic.feature.session
 
 import com.wire.kalium.logic.NetworkFailure
-import com.wire.kalium.logic.data.event.EventRepository
+import com.wire.kalium.logic.data.auth.login.LoginRepository
 import com.wire.kalium.logic.functional.Either
 import com.wire.kalium.logic.test_util.TestNetworkException
-import com.wire.kalium.network.api.notification.pushToken.PushTokenRequestBody
+import com.wire.kalium.network.api.user.pushToken.PushTokenRequestBody
 import io.mockative.Mock
 import io.mockative.classOf
 import io.mockative.given
@@ -22,19 +22,19 @@ import kotlin.test.assertIs
 class PushTokenUseCaseTest {
 
     @Mock
-    val eventRepository = mock(classOf<EventRepository>())
+    val loginRepository = mock(classOf<LoginRepository>())
 
     private lateinit var pushTokenUseCase: RegisterTokenUseCase
 
     @BeforeTest
     fun setup() {
-        pushTokenUseCase = RegisterTokenUseCase(eventRepository)
+        pushTokenUseCase = RegisterTokenUseCase(loginRepository)
     }
 
 
     @Test
     fun givenRepositoryCallIsSuccessful_thenSuccessIsReturned() = runTest {
-        given(eventRepository)
+        given(loginRepository)
             .coroutine { registerToken(pushTokenRequestBody) }
             .then { Either.Right(Unit) }
 
@@ -45,7 +45,7 @@ class PushTokenUseCaseTest {
 
         assertIs<RegisterTokenResult.Success>(actual)
 
-        verify(eventRepository)
+        verify(loginRepository)
             .coroutine { registerToken(pushTokenRequestBody) }
             .wasInvoked(exactly = once)
     }
@@ -55,7 +55,7 @@ class PushTokenUseCaseTest {
     fun givenRepositoryCallFailWithAppNotFound_thenInvalidCodeIsReturned() = runTest {
         val expected = NetworkFailure.ServerMiscommunication(TestNetworkException.invalidCode)
 
-        given(eventRepository)
+        given(loginRepository)
             .coroutine { registerToken(pushTokenRequestBody) }
             .then { Either.Left(expected) }
 
@@ -66,7 +66,7 @@ class PushTokenUseCaseTest {
 
         assertIs<RegisterTokenResult.Failure.AppNotFound>(actual)
 
-        verify(eventRepository)
+        verify(loginRepository)
             .coroutine { registerToken(pushTokenRequestBody) }
             .wasInvoked(exactly = once)
     }
@@ -75,7 +75,7 @@ class PushTokenUseCaseTest {
     fun givenRepositoryCallFail_thenErrorIsReturned() = runTest {
         val expected = NetworkFailure.ServerMiscommunication(TestNetworkException.generic)
 
-        given(eventRepository)
+        given(loginRepository)
             .coroutine { registerToken(pushTokenRequestBody) }
             .then { Either.Left(expected) }
 
@@ -87,7 +87,7 @@ class PushTokenUseCaseTest {
         assertIs<RegisterTokenResult.Failure.Generic>(actual)
         assertEquals(expected, actual.failure)
 
-        verify(eventRepository)
+        verify(loginRepository)
             .coroutine { registerToken(pushTokenRequestBody) }
             .wasInvoked(exactly = once)
     }
