+++ network/src/commonTest/kotlin/com/wire/kalium/api/tools/json/api/notification/NotificationApiTest.kt
@@ -11,7 +11,6 @@ import io.ktor.http.HttpStatusCode
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
 import kotlin.test.Test
-import kotlin.test.assertEquals
 import kotlin.test.assertIs
 import kotlin.test.assertTrue
 
@@ -94,23 +93,6 @@ class NotificationApiTest : ApiTest {
     }
 
     @Test
-    fun givenAValidRequest_whenRegisteredValidToken_theTokenRegisteredSuccessfully() = runTest {
-        val httpClient = mockAuthenticatedHttpClient(
-            NotificationEventsResponseJson.registerTokenResponse, statusCode = HttpStatusCode.Created, assertion = {
-                assertPost()
-                assertBodyContent(VALID_PUSH_TOKEN_REQUEST.rawJson)
-            }
-        )
-        val notificationsApi = NotificationApiImpl(httpClient, TEST_BACKEND_CONFIG)
-
-        notificationsApi.registerToken(VALID_PUSH_TOKEN_REQUEST.serializableData)
-
-        val actual = notificationsApi.registerToken(VALID_PUSH_TOKEN_REQUEST.serializableData)
-        assertIs<NetworkResponse.Success<Unit>>(actual)
-        assertTrue(actual.isSuccessful())
-    }
-
-    @Test
     fun given404Response_whenGettingNotificationGettingAllNotifications_thenTheResponseIsParsedCorrectly() = runTest {
         val httpClient = mockAuthenticatedHttpClient(
             NotificationEventsResponseJson.notificationsWithUnknownEventAtFirstPosition,
@@ -125,7 +107,6 @@ class NotificationApiTest : ApiTest {
     }
 
     private companion object {
-        val VALID_PUSH_TOKEN_REQUEST = NotificationEventsResponseJson.validPushTokenRequest
         const val PATH_NOTIFICATIONS = "/notifications"
         const val SIZE_QUERY_KEY = "size"
         const val CLIENT_QUERY_KEY = "client"
