+++ network/src/commonTest/kotlin/com/wire/kalium/api/tools/json/api/user/login/LoginApiTest.kt
@@ -10,6 +10,7 @@ import com.wire.kalium.network.api.model.UserDTO
 import com.wire.kalium.network.api.user.login.LoginApi
 import com.wire.kalium.network.api.user.login.LoginApiImpl
 import com.wire.kalium.network.exceptions.KaliumException
+import com.wire.kalium.network.utils.NetworkResponse
 import com.wire.kalium.network.utils.isSuccessful
 import io.ktor.http.HttpStatusCode
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -17,6 +18,7 @@ import kotlinx.coroutines.test.runTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
+import kotlin.test.assertIs
 import kotlin.test.assertTrue
 
 @ExperimentalCoroutinesApi
@@ -105,6 +107,23 @@ class LoginApiTest : ApiTest {
 
     }
 
+    @Test
+    fun givenAValidRequest_whenRegisteredValidToken_theTokenRegisteredSuccessfully() = runTest {
+        val httpClient = mockAuthenticatedHttpClient(
+            RegisterTokenJson.registerTokenResponse, statusCode = HttpStatusCode.Created, assertion = {
+                assertPost()
+                assertBodyContent(VALID_PUSH_TOKEN_REQUEST.rawJson)
+            }
+        )
+        val loginApi = LoginApiImpl(httpClient)
+
+        loginApi.registerToken(VALID_PUSH_TOKEN_REQUEST.serializableData)
+
+        val actual = loginApi.registerToken(VALID_PUSH_TOKEN_REQUEST.serializableData)
+        assertIs<NetworkResponse.Success<Unit>>(actual)
+        assertTrue(actual.isSuccessful())
+    }
+
     private companion object {
         val refreshToken = "415a5306-a476-41bc-af36-94ab075fd881"
         val userID = QualifiedID("user_id", "user.domain.io")
@@ -141,5 +160,7 @@ class LoginApiTest : ApiTest {
         const val PATH_LOGIN = "/login"
         const val PATH_SELF = "/self"
         const val TEST_HOST = """https://test-https.wire.com"""
+        val VALID_PUSH_TOKEN_REQUEST = RegisterTokenJson.validPushTokenRequest
+
     }
 }
